making the correct changes
